moved out to a hook-like function
